+++ src/DiscogsApiClient.Tests/Usings.cs
@@ -1,6 +1,7 @@
 global using DiscogsApiClient.Authentication;
 global using DiscogsApiClient.Contract;
 global using DiscogsApiClient.Contract.User;
+global using DiscogsApiClient.Contract.User.List;
 global using DiscogsApiClient.Exceptions;
 global using DiscogsApiClient.QueryParameters;
 global using DiscogsApiClient.Tests.Fixtures;